+++ test/sync/src/androidTest/kotlin/io/realm/test/mongodb/shared/CredentialsTests.kt
@@ -18,8 +18,10 @@ package io.realm.test.mongodb.shared
 
 import io.realm.internal.platform.runBlocking
 import io.realm.mongodb.App
+import io.realm.mongodb.AppException
 import io.realm.mongodb.AuthenticationProvider
 import io.realm.mongodb.Credentials
+import io.realm.mongodb.GoogleAuthType
 import io.realm.test.mongodb.TestApp
 import io.realm.test.mongodb.asTestApp
 import io.realm.test.mongodb.createUserAndLogIn
@@ -29,6 +31,8 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
 import kotlin.test.assertNotNull
+import kotlin.test.assertTrue
+import kotlin.test.fail
 
 @Suppress("ForbiddenComment")
 class CredentialsTests {
@@ -56,55 +60,64 @@ class CredentialsTests {
         }
     }
 
+    @Test
+    fun allCredentials_emptyInputThrows() {
+        for (value in AuthenticationProvider.values()) {
+            assertFailsWith<IllegalArgumentException> { // No arguments should be allow
+                when (value) {
+                    AuthenticationProvider.ANONYMOUS -> { /* Do nothing, no arguments */ }
+                    AuthenticationProvider.API_KEY -> Credentials.apiKey("")
+                    AuthenticationProvider.APPLE -> Credentials.apple("")
+                    AuthenticationProvider.EMAIL_PASSWORD -> { /* Test below as a special case */ }
+                    AuthenticationProvider.FACEBOOK -> Credentials.facebook("")
+                    AuthenticationProvider.GOOGLE -> { /* Tested below as a special case */ }
+                    AuthenticationProvider.JWT -> Credentials.jwt("")
+                }
+            }
+        }
+
+        // Test Email/Password as a special case, due to it having two arguments
+        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("", "password") }
+        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("foo@bar.com", "") }
+
+        // Test Google as a special case as two types of Google login exists
+        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.AUTH_CODE) }
+        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.ID_TOKEN) }
+    }
+
     fun anonymous(): Credentials {
-        val creds = Credentials.anonymous()
+        val creds: Credentials = Credentials.anonymous()
         // assertTrue(creds.asJson().contains("anon-user")) // Treat the JSON as an opaque value.
         return creds
     }
 
-//    @Test
-//    fun apiKey() {
-//        val creds = Credentials.apiKey("token")
-//        assertEquals(Credentials.Provider.API_KEY, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("token")) // Treat the JSON as an opaque value.
-//    }
-//
-//    @Test
-//    fun apiKey_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.apiKey("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.apiKey(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun apple() {
-//        val creds = Credentials.apple("apple-token")
-//        assertEquals(Credentials.Provider.APPLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("apple-token")) // Treat the JSON as a largely opaque value.
-//    }
-//
-//    @Test
-//    fun apple_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.apple("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.apple(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun customFunction() {
-//        val mail = TestHelper.getRandomEmail()
-//        val id = 666 + TestHelper.getRandomId()
-//        val creds = mapOf(
-//            "mail" to mail,
-//            "id" to id
-//        ).let { Credentials.customFunction(Document(it)) }
-//        assertEquals(Credentials.Provider.CUSTOM_FUNCTION, creds.identityProvider)
-//        assertTrue(creds.asJson().contains(mail))
-//        assertTrue(creds.asJson().contains(id.toString()))
-//    }
-//
-//    @Test
-//    fun customFunction_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.customFunction(null) }
-//    }
+    @Test
+    fun apiKey() {
+       val creds: Credentials = Credentials.apiKey("token")
+       assertEquals(AuthenticationProvider.API_KEY, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("token")) // Treat the JSON as an opaque value.
+    }
+
+    @Test
+    fun apple() {
+       val creds = Credentials.apple("apple-token")
+       assertEquals(AuthenticationProvider.APPLE, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("apple-token")) // Treat the JSON as a largely opaque value.
+    }
+
+    // TODO See https://github.com/realm/realm-kotlin/issues/742
+    // @Test
+    // fun customFunction() {
+    //     val mail = TestHelper.getRandomEmail()
+    //     val id = 666 + TestHelper.getRandomId()
+    //     val creds = mapOf(
+    //         "mail" to mail,
+    //         "id" to id
+    //     ).let { Credentials.customFunction(Document(it)) }
+    //     assertEquals(Credentials.Provider.CUSTOM_FUNCTION, creds.identityProvider)
+    //     assertTrue(creds.asJson().contains(mail))
+    //     assertTrue(creds.asJson().contains(id.toString()))
+    // }
 
     fun emailPassword(): Credentials {
         val creds = Credentials.emailPassword("foo@bar.com", "secret")
@@ -117,68 +130,38 @@ class CredentialsTests {
     }
 
     @Test
-    fun emailPassword_invalidInput() {
-        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("", "password") }
-        assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("email", "") }
-        // TODO I guess we cannot workaround these being null in Kotlin, but do we need some Java tests then?
-        // assertFailsWith<IllegalArgumentException> { Credentials.emailPassword(TestHelper.getNull(), "password") }
-        // assertFailsWith<IllegalArgumentException> { Credentials.emailPassword("email", TestHelper.getNull()) }
+    fun facebook() {
+       val creds = Credentials.facebook("fb-token")
+       assertEquals(AuthenticationProvider.FACEBOOK, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("fb-token"))
     }
 
-//    @Test
-//    fun facebook() {
-//        val creds = Credentials.facebook("fb-token")
-//        assertEquals(Credentials.Provider.FACEBOOK, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("fb-token"))
-//    }
-//
-//    @Test
-//    fun facebook_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.facebook("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.facebook(TestHelper.getNull()) }
-//    }
-//
-//    @Test
-//    fun google_authCode() {
-//        val creds = Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
-//        assertEquals(Credentials.Provider.GOOGLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("google-token"))
-//        assertTrue(creds.asJson().contains("authCode"))
-//    }
-//
-//    @Test
-//    fun google_idToken() {
-//        val creds = Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
-//        assertEquals(Credentials.Provider.GOOGLE, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("google-token"))
-//        assertTrue(creds.asJson().contains("id_token"))
-//    }
-//
-//    @Test
-//    fun google_invalidInput_authCode() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.AUTH_CODE) }
-//        assertFailsWith<IllegalArgumentException> { Credentials.google(TestHelper.getNull(), GoogleAuthType.AUTH_CODE) }
-//    }
-//
-//    @Test
-//    fun google_invalidInput_idToken() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.google("", GoogleAuthType.ID_TOKEN) }
-//        assertFailsWith<IllegalArgumentException> { Credentials.google(TestHelper.getNull(), GoogleAuthType.ID_TOKEN) }
-//    }
-//
-//    @Ignore("FIXME: Awaiting ObjectStore support")
-//    @Test
-//    fun jwt() {
-//        val creds = Credentials.jwt("jwt-token")
-//        assertEquals(Credentials.Provider.JWT, creds.identityProvider)
-//        assertTrue(creds.asJson().contains("jwt-token"))
-//    }
-//
-//    @Test
-//    fun jwt_invalidInput() {
-//        assertFailsWith<IllegalArgumentException> { Credentials.jwt("") }
-//        assertFailsWith<IllegalArgumentException> { Credentials.jwt(TestHelper.getNull()) }
-//    }
+    @Test
+    fun google_authCode() {
+        // TODO https://github.com/realm/realm-core/issues/5347
+        assertFailsWith<NotImplementedError> {
+            Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
+        }
+       // val creds = Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
+       // assertEquals(AuthenticationProvider.GOOGLE, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("google-token"))
+       // assertTrue(creds.asJson().contains("authCode"))
+    }
+
+    @Test
+    fun google_idToken() {
+       val creds = Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
+       assertEquals(AuthenticationProvider.GOOGLE, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("google-token"))
+       // assertTrue(creds.asJson().contains("id_token"))
+    }
+
+    @Test
+    fun jwt() {
+       val creds = Credentials.jwt("jwt-token")
+       assertEquals(AuthenticationProvider.JWT, creds.authenticationProvider)
+       // assertTrue(creds.asJson().contains("jwt-token"))
+    }
 
     @Test
     fun loginUsingCredentials() {
@@ -190,6 +173,8 @@ class CredentialsTests {
                         val user = app.login(Credentials.anonymous())
                         assertNotNull(user)
                     }
+// TODO Enable this when https://github.com/realm/realm-kotlin/issues/432 is complete.
+
 //                    AuthenticationProvider.API_KEY -> {
 //                        // Log in, create an API key, log out, log in with the key, compare users
 //                        val user: User =
@@ -199,6 +184,9 @@ class CredentialsTests {
 //                        val apiKeyUser = app.login(Credentials.apiKey(key.value!!))
 //                        assertEquals(user.id, apiKeyUser.id)
 //                    }
+
+// TODO Enable this when https://github.com/realm/realm-kotlin/issues/741 is complete.
+
 //                    Credentials.Provider.CUSTOM_FUNCTION -> {
 //                        val customFunction = mapOf(
 //                            "mail" to TestHelper.getRandomEmail(),
@@ -217,47 +205,23 @@ class CredentialsTests {
                         val user = app.createUserAndLogIn(email, password)
                         assertNotNull(user)
                     }
-
-//                    // These providers are hard to test for real since they depend on a 3rd party
-//                    // login service. Instead we attempt to login and verify that a proper exception
-//                    // is thrown. At least that should verify that correctly formatted JSON is being
-//                    // sent across the wire.
-//                    AuthenticationProvider.FACEBOOK -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.facebook("facebook-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.APPLE -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.apple("apple-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.GOOGLE -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.google("google-token", GoogleAuthType.AUTH_CODE)
-//                        )
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.google("google-token", GoogleAuthType.ID_TOKEN)
-//                        )
-//                    }
-//                    AuthenticationProvider.JWT -> {
-//                        expectErrorCode(
-//                            app,
-//                            ErrorCode.INVALID_SESSION,
-//                            Credentials.jwt("jwt-token")
-//                        )
-//                    }
-//                    AuthenticationProvider.UNKNOWN -> {
-//                        // Ignore
-//                    }
+                    // These providers are hard to test for real since they depend on a 3rd party
+                    // login service. Instead we attempt to login and verify that a proper exception
+                    // is thrown. At least that should verify that correctly formatted JSON is being
+                    // sent across the wire.
+                    AuthenticationProvider.FACEBOOK -> {
+                       expectInvalidSession(app, Credentials.facebook("facebook-token"))
+                    }
+                    AuthenticationProvider.APPLE -> {
+                       expectInvalidSession(app, Credentials.apple("apple-token"))
+                    }
+                    AuthenticationProvider.GOOGLE -> {
+                       expectInvalidSession(app, Credentials.google("google-token", GoogleAuthType.AUTH_CODE))
+                       expectInvalidSession(app, Credentials.google("google-token", GoogleAuthType.ID_TOKEN))
+                    }
+                    AuthenticationProvider.JWT -> {
+                       expectInvalidSession(app, Credentials.jwt("jwt-token"))
+                    }
                     else -> {
                         error("Untested provider: $provider")
                     }
@@ -266,12 +230,14 @@ class CredentialsTests {
         }
     }
 
-//    private fun expectErrorCode(app: App, expectedCode: ErrorCode, credentials: Credentials) {
-//        try {
-//            app.login(credentials)
-//            fail()
-//        } catch (error: AppException) {
-//            assertEquals(expectedCode, error.errorCode)
-//        }
-//    }
+    private fun expectInvalidSession(app: App, credentials: Credentials) {
+        try {
+            runBlocking {
+                app.login(credentials)
+            }
+            fail()
+        } catch (error: AppException) {
+            assertTrue(error.message!!.contains("BoOM"), error.message)
+        }
+    }
 }